vision/pyproject.toml
[project]
name = "historicam-vision"
version = "0.1.0"
description = "Vision pipeline for HistoriCam - embedding generation and vector search"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-aiplatform>=1.70.0",
    "google-cloud-storage>=2.18.2",
    "pandas>=2.2.3",
    "numpy>=2.1.3",
    "pillow>=11.0.0",
    "vertexai>=1.43.0",
]

# gcs_utils is mounted at runtime via volume mount
# Install at runtime with: uv pip install -e /app/lib/gcs_utils

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
